fix: mlflow entities mapping pt. 1 -- READ MESSAGE
This commit cleans the migration history, and it will require to
recreate the database from fresh, or truncate all it's tables and rerun
the `app/prestart.sh` script. I had to clean the history because to
write the migration would be longer than what I thought it was worth it,
and the `alembic revision --autogenerate` command was not creating a
runnable migration script due to the tool's limitation dealing with the number
of schema changes applied at once.

Besides that, this commit adjusts foreign keys and table fields to
properly support mlflow entity mappings and fix hidden problems due to
required fields missing the `nullable=False` constraint in their table.

A `mlflow_<something>` column was added to the mapped entities.
The Experiment entity can be mapepd through it's mlflow_id attribute
The Model can be mapped through it's mlflow_name attribute
The ModelVersion entity can be mapped through it's parent Model
mlflow_name and it's own mlflow_name.

This commit is a WIP since it only adjusts the entity scripts,
but several fixes need to be done in the repository DTOs and possibly
controller implementations.