fix: error due model creation request missing "bastch" in GlobalPooling forwardArgs
The positional layer with Optional typing was being interpreted as a
required layer by the code generator. the problem was solved by making
it a keyword argument with default value None.
It could also be solved by allowing frontend to reference the
batch array DataModule created during training, but it would be harder
and we can fill this parameter automatically for the user.